+++ src/components/PostUploader/CommentUploader.scss
@@ -1,0 +1,9 @@
+@import '../../stylesheets/util.scss';
+
+.comment-uploader {
+  margin: 4rem 0;
+  width: 60rem;
+  height: 10rem;
+  font-size: 3rem;
+  border: 1px solid $gray-border-color;
+}
