+++ src/components/PostUploader/ImageUploader.scss
@@ -1,0 +1,11 @@
+@import '../../stylesheets/util.scss';
+
+.image-uploader {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  margin: 4rem 0;
+  width: 60rem;
+  height: 40rem;
+  border: 1px solid $gray-border-color;
+}
