+++ src/components/PostUploader/LocationUploader.jsx
@@ -1,0 +1,11 @@
+import React from 'react';
+
+const LocationUploader = () => {
+  return (
+    <div>
+      <button>장소검색</button>
+    </div>
+  );
+};
+
+export default LocationUploader;
